fix: cast X_WSAError to uint32_t for comparison in xam_net.cc

+++ src/xenia/kernel/xam/xam_net.cc
@@ -531,7 +531,8 @@ dword_result_t NetDll_WSARecvFrom_entry(
                               from_ptr, fromlen_ptr, overlapped_ptr);
     if (ret < 0) {
       auto err = socket->GetLastWSAError();
-      if (err == X_WSAError::X_WSAEWOULDBLOCK || err == X_WSAError::X_WSA_IO_PENDING) {
+      if (static_cast<uint32_t>(err) == static_cast<uint32_t>(X_WSAError::X_WSAEWOULDBLOCK) || 
+          static_cast<uint32_t>(err) == static_cast<uint32_t>(X_WSAError::X_WSA_IO_PENDING)) {
         // Retry on non-fatal errors
         std::this_thread::sleep_for(std::chrono::milliseconds(100));
         continue;
@@ -612,7 +613,8 @@ dword_result_t NetDll_WSASendTo_entry(
           std::this_thread::sleep_for(std::chrono::milliseconds(100));
           continue;
         }
-        if (err == X_WSAError::X_WSAEWOULDBLOCK || err == X_WSAError::X_WSA_IO_PENDING) {
+        if (static_cast<uint32_t>(err) == static_cast<uint32_t>(X_WSAError::X_WSAEWOULDBLOCK) || 
+            static_cast<uint32_t>(err) == static_cast<uint32_t>(X_WSAError::X_WSA_IO_PENDING)) {
           XThread::SetLastError(err);
         } else {
           XThread::SetLastError(0);
